breaking up search into multiple files and a script
